WIP on rewind

+++ app/src/main/java/org/kpcc/android/LiveFragment.java
@@ -3,6 +3,7 @@ package org.kpcc.android;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
+import android.os.SystemClock;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -167,7 +168,7 @@ public class LiveFragment extends Fragment {
             @Override
             public void onClick(View v) {
                 if (streamBundle != null) {
-                    int currentPosition = streamBundle.liveStream.getCurrentPosition();
+                    long currentPosition = streamBundle.liveStream.getCurrentPosition();
                     streamBundle.liveStream.seekTo(currentPosition - 1000*30);
                 }
             }
